+++ lib/coal.dart
@@ -1,4 +1,5 @@
 export 'args.dart';
 export 'keypass.dart';
+export 'readline.dart';
 export 'tab.dart';
 export 'utils.dart';